ErrorControl: Match functions to prototypes in header.

+++ OERRCTRL.cpp
@@ -37,7 +37,7 @@ void ErrorControl::deinit()
 {
 }
 
-void ErrorControl::set_dp_id(char ecPlayerId, DWORD dpPlayerId )
+void ErrorControl::set_dp_id(char ecPlayerId, long unsigned int dpPlayerId )
 {
 	if( ecPlayerId != self_ec_player_id )
 	{
@@ -58,7 +58,7 @@ void ErrorControl::set_dp_id(char ecPlayerId, DWORD dpPlayerId )
 }
 
 // return ec_player_id, 0 for not found (can't found own dpPlayerId)
-char ErrorControl::get_ec_player_id( DWORD dpPlayerId )
+char ErrorControl::get_ec_player_id( long unsigned int dpPlayerId )
 {
 	if( dpPlayerId == BROADCAST_PID || dpPlayerId == 0)
 		return 0;
@@ -71,7 +71,7 @@ char ErrorControl::get_ec_player_id( DWORD dpPlayerId )
 }
 
 // return 1 on success, -1 if queue is_full, 0 for other failure
-int ErrorControl::send(char ecPlayerId, void *dataPtr, DWORD dataLen)
+int ErrorControl::send(char ecPlayerId, void *dataPtr, long unsigned int dataLen)
 {
 	if( connecting_player_count == 0)
 		return 1;
@@ -152,7 +152,7 @@ int ErrorControl::send(char ecPlayerId, void *dataPtr, DWORD dataLen)
 	return 1;
 }
 
-char *ErrorControl::receive(char *sendEcPlayerId, DWORD *dataLen)
+char *ErrorControl::receive(char *sendEcPlayerId, long unsigned int *dataLen)
 {
 	// ----- draw the head of recv_queue ----- //
 
@@ -551,7 +551,7 @@ int ErrorControl::recv_queue_space()
 }
 
 
-void ErrorControl::en_recv_queue(void *dataPtr, DWORD dataLen)
+void ErrorControl::en_recv_queue(void *dataPtr, long unsigned int dataLen)
 {
 	if( is_recv_full() )
 	{
@@ -590,7 +590,7 @@ void ErrorControl::clear_ack(char frameId)
 	memset( ack_flag[frameId], 0, MAX_PLAYER );
 }
 
-void ErrorControl::mark_send_time(char frameId, DWORD duration)
+void ErrorControl::mark_send_time(char frameId, long unsigned int duration)
 {
 	send_time[frameId] = m.get_time();
 	re_send_after[frameId] = duration;
